Changelog PoC

+++ src/models.rs
@@ -5,6 +5,7 @@ use petgraph::visit::{Dfs, IntoNeighborsDirected, NodeCount};
 use petgraph::Direction;
 use rusqlite::types::Value;
 use rusqlite::{params_from_iter, Connection};
+use sha2::{Digest, Sha256};
 use std::collections::{HashMap, HashSet};
 use std::fmt::*;
 use std::hash::Hash;
@@ -14,11 +15,11 @@ pub mod edge;
 pub mod path;
 pub mod sequence;
 
-use crate::models;
 use crate::models::block::Block;
 use crate::models::edge::Edge;
 use crate::models::path::{all_simple_paths, Path, PathBlock};
 use crate::models::sequence::Sequence;
+use crate::{get_overlap, models};
 
 #[derive(Debug)]
 pub struct Collection {
@@ -66,12 +67,22 @@ pub struct Sample {
 impl Sample {
     pub fn create(conn: &mut Connection, name: &String) -> Sample {
         let mut stmt = conn
-            .prepare("INSERT INTO sample (name) VALUES (?1) RETURNING *")
+            .prepare("INSERT INTO sample (name) VALUES (?1)")
             .unwrap();
-        let mut rows = stmt
-            .query_map((name,), |row| Ok(Sample { name: row.get(0)? }))
-            .unwrap();
-        rows.next().unwrap().unwrap()
+        match stmt.execute((name,)) {
+            Ok(_) => Sample { name: name.clone() },
+            Err(rusqlite::Error::SqliteFailure(err, details)) => {
+                if err.code == rusqlite::ErrorCode::ConstraintViolation {
+                    println!("{err:?} {details:?}");
+                    Sample { name: name.clone() }
+                } else {
+                    panic!("something bad happened querying the database")
+                }
+            }
+            Err(_) => {
+                panic!("something bad happened querying the database")
+            }
+        }
     }
 }
 
@@ -100,7 +111,7 @@ impl BlockGroup {
                 name: row.get(3)?,
             })
         }) {
-            Ok(path) => path,
+            Ok(res) => res,
             Err(rusqlite::Error::SqliteFailure(err, details)) => {
                 if err.code == rusqlite::ErrorCode::ConstraintViolation {
                     println!("{err:?} {details:?}");
@@ -331,11 +342,23 @@ impl BlockGroup {
         path_id: i32,
         start: i32,
         end: i32,
-        new_block_id: i32,
+        new_block: &Block,
         chromosome_index: i32,
         phased: i32,
     ) {
-        println!("change is {path_id} {start} {end} {new_block_id}");
+        let new_block_id = new_block.id;
+        let change = ChangeLog::new(
+            path_id,
+            start,
+            end,
+            new_block.sequence_hash.clone(),
+            new_block.start,
+            new_block.end,
+            new_block.strand.clone(),
+        );
+        if ChangeLog::exists(conn, &change.hash) {
+            return;
+        }
         // todo:
         // 1. get blocks where start-> end overlap
         // 2. split old blocks at boundary points, make new block for left/right side
@@ -354,7 +377,6 @@ impl BlockGroup {
 
         let path = Path::get(conn, path_id);
         let graph = PathBlock::blocks_to_graph(conn, path.id);
-        println!("{path:?} {graph:?}");
         let query = format!("SELECT id, sequence_hash, block_group_id, start, end, strand from block where id in ({block_ids})", block_ids = graph.nodes().map(|k| format!("{k}")).collect::<Vec<_>>().join(","));
         let mut stmt = conn.prepare(&query).unwrap();
         let mut blocks: HashMap<i32, Block> = HashMap::new();
@@ -381,14 +403,22 @@ impl BlockGroup {
         let mut path_end = 0;
         let mut new_edges = vec![];
         let mut previous_block: Option<&Block> = None;
+        let mut loose_connection = false;
         for block_id in &path.blocks {
             let block = blocks.get(block_id).unwrap();
             let block_length = (block.end - block.start);
             path_end += block_length;
 
-            let contains_start = path_start <= start && start < path_end;
-            let contains_end = path_start <= end && end < path_end;
-            let overlap = path_start <= end && start <= path_end;
+            if loose_connection && path_start == end {
+                new_edges.push((Some(new_block_id), Some(block.id)));
+                loose_connection = false;
+            }
+
+            let (contains_start, contains_end, overlap) =
+                get_overlap(path_start, path_end, start, end);
+            println!(
+                "{path_start} {path_end} {start} {end} {contains_start} {contains_end} {overlap}"
+            );
 
             if contains_start && contains_end {
                 // our range is fully contained w/in the block
@@ -428,15 +458,22 @@ impl BlockGroup {
                 // our range is overlapping the end of the block
                 // |----block---|
                 //        |----range---|
-                let (left_block, right_block) = Block::split(
-                    conn,
-                    block,
-                    block.start + start - path_start,
-                    chromosome_index,
-                    phased,
-                )
-                .unwrap();
-                Block::delete(conn, block.id);
+                let split_point = block.start + start - path_start;
+                println!("{block:?} {split_point}");
+                if split_point == block.start {
+                    if let Some(pb) = previous_block {
+                        new_edges.push((Some(pb.id), Some(new_block_id)));
+                    }
+                    // we actually are skipping this block
+                    if path_end >= end {
+                        loose_connection = true;
+                    }
+                } else {
+                    let (left_block, right_block) =
+                        Block::split(conn, block, split_point, chromosome_index, phased).unwrap();
+                    Block::delete(conn, block.id);
+                    new_edges.push((Some(left_block.id), Some(new_block_id)));
+                }
                 // let left_block = Block::create(
                 //     conn,
                 //     &block.sequence_hash,
@@ -450,37 +487,15 @@ impl BlockGroup {
                 // } else {
                 //     new_edges.push((None, Some(left_block.id)));
                 // }
-                new_edges.push((Some(left_block.id), Some(new_block_id)));
             } else if contains_end {
                 // our range is overlapping the beginning of the block
                 //              |----block---|
                 //        |----range---|
-                let (left_block, right_block) = Block::split(
-                    conn,
-                    block,
-                    block.start + end - path_start,
-                    chromosome_index,
-                    phased,
-                )
-                .unwrap();
+                let split_point = block.start + end - path_start;
+                let (left_block, right_block) =
+                    Block::split(conn, block, split_point, chromosome_index, phased).unwrap();
                 Block::delete(conn, block.id);
-                // let right_block = Block::create(
-                //     conn,
-                //     &block.sequence_hash,
-                //     block_group_id,
-                //     end - path_start,
-                //     block.end,
-                //     &block.strand,
-                // );
-                // // what stuff went to this block?
                 new_edges.push((Some(new_block_id), Some(right_block.id)));
-                // let last_node = dfs.next(&graph);
-                // if last_node.is_some() {
-                //     let next_block = blocks.get(&(last_node.unwrap() as i32)).unwrap();
-                //     new_edges.push((Some(right_block.id), Some(next_block.id)));
-                // } else {
-                //     new_edges.push((Some(right_block.id), None))
-                // }
                 break;
             } else if overlap {
                 // our range is the whole block, ignore it
@@ -500,6 +515,79 @@ impl BlockGroup {
         for new_edge in new_edges {
             Edge::create(conn, new_edge.0, new_edge.1, chromosome_index, phased);
         }
+
+        change.save(conn);
+    }
+}
+
+pub struct ChangeLog {
+    hash: String,
+    path_id: i32,
+    path_start: i32,
+    path_end: i32,
+    seq_hash: String,
+    seq_start: i32,
+    seq_end: i32,
+    strand: String,
+}
+
+impl ChangeLog {
+    pub fn new(
+        path_id: i32,
+        path_start: i32,
+        path_end: i32,
+        seq_hash: String,
+        seq_start: i32,
+        seq_end: i32,
+        seq_strand: String,
+    ) -> ChangeLog {
+        let mut hasher = Sha256::new();
+        hasher.update(path_id.to_string());
+        hasher.update(path_start.to_string());
+        hasher.update(path_end.to_string());
+        hasher.update(&seq_hash);
+        hasher.update(seq_start.to_string());
+        hasher.update(seq_end.to_string());
+        hasher.update(&seq_strand);
+        let result = hasher.finalize();
+        let hash = format!("{:x}", result);
+        ChangeLog {
+            hash,
+            path_id,
+            path_start,
+            path_end,
+            seq_hash,
+            seq_start,
+            seq_end,
+            strand: seq_strand,
+        }
+    }
+
+    pub fn save(&self, conn: &Connection) {
+        ChangeLog::create(conn, self);
+    }
+
+    pub fn create(conn: &Connection, change_log: &ChangeLog) {
+        let mut stmt = conn
+            .prepare("INSERT INTO change_log (hash, path_id, path_start, path_end, sequence_hash, sequence_start, sequence_end, sequence_strand) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8);")
+            .unwrap();
+        let placeholders = vec![
+            Value::from(change_log.hash.clone()),
+            Value::from(change_log.path_id),
+            Value::from(change_log.path_start),
+            Value::from(change_log.path_end),
+            Value::from(change_log.seq_hash.clone()),
+            Value::from(change_log.seq_start),
+            Value::from(change_log.seq_end),
+            Value::from(change_log.strand.clone()),
+        ];
+        stmt.execute(params_from_iter(placeholders)).unwrap();
+    }
+
+    pub fn exists(conn: &mut Connection, hash: &String) -> bool {
+        let query = "SELECT hash from change_log where hash = ?1;";
+        let mut stmt = conn.prepare(query).unwrap();
+        stmt.exists((hash,)).unwrap()
     }
 }
 
@@ -544,9 +632,8 @@ mod tests {
     }
 
     #[test]
-    fn simple_insert() {
-        fs::remove_file("test.db");
-        let mut conn = get_db_connection("test.db");
+    fn insert_and_deletion() {
+        let mut conn = get_connection();
         let (block_group_id, path_id) = setup_block_group(&mut conn);
         let insert_sequence =
             Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
@@ -558,7 +645,7 @@ mod tests {
             4,
             &"1".to_string(),
         );
-        BlockGroup::insert_change(&mut conn, path_id, 7, 15, insert.id, 1, 0);
+        BlockGroup::insert_change(&mut conn, path_id, 7, 15, &insert, 1, 0);
 
         let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
         assert_eq!(
@@ -582,7 +669,7 @@ mod tests {
         );
 
         // take out an entire block.
-        BlockGroup::insert_change(&mut conn, path_id, 19, 31, deletion.id, 1, 0);
+        BlockGroup::insert_change(&mut conn, path_id, 19, 31, &deletion, 1, 0);
         let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
         assert_eq!(
             all_sequences,
@@ -593,5 +680,199 @@ mod tests {
                 "AAAAAAANNNNTTTTGGGGGGGGG".to_string(),
             ])
         )
+    }
+
+    #[test]
+    fn simple_insert() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 7, 15, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_on_block_boundary_start() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 10, 10, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAAANNNNTTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_on_block_boundary_end() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 9, 9, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAANNNNATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_across_entire_block_boundary() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 10, 20, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAAANNNNCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_across_two_blocks() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 15, 25, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAAATTTTTNNNNCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn simple_deletion() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let deletion_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"".to_string(), true);
+        let deletion = Block::create(
+            &conn,
+            &deletion_sequence,
+            block_group_id,
+            0,
+            0,
+            &"1".to_string(),
+        );
+
+        // take out an entire block.
+        BlockGroup::insert_change(&mut conn, path_id, 19, 31, &deletion, 1, 0);
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAAATTTTTTTTTGGGGGGGGG".to_string(),
+            ])
+        )
+    }
+
+    #[test]
+    fn doesnt_apply_same_insert_twice() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 7, 15, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+
+        BlockGroup::insert_change(&mut conn, path_id, 7, 15, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
     }
 }
